Moved method to NotificationeService
