+++ service-api/src/main/java/greencity/service/notification/NotificationeService.java
@@ -1,10 +1,11 @@
 package greencity.service.notification;
 
+import greencity.dto.AddingViolationsToUserDto;
 import greencity.dto.NotificationTemplateDto;
 import greencity.dto.PageableDto;
 import org.springframework.data.domain.Pageable;
 
-public interface ManagementNotificationService {
+public interface NotificationeService {
     /**
      * Method that update notification template.
      *
@@ -25,4 +26,12 @@ public interface ManagementNotificationService {
      * @author Dima Sannytski
      */
     NotificationTemplateDto findById(Long id);
+
+    /**
+     * Method for send email with description to user.
+     *
+     * @param dto {@link AddingViolationsToUserDto } order id with description.
+     * @author Veremchuk Zakhar.
+     */
+    void sendNotificationAboutViolation(AddingViolationsToUserDto dto, String language);
 }
