+++ service/src/main/java/greencity/service/notification/NotificationeServiceImpl.java
@@ -1,15 +1,16 @@
 package greencity.service.notification;
 
+import greencity.client.RestClient;
 import greencity.constant.ErrorMessage;
-import greencity.dto.NotificationScheduleDto;
-import greencity.dto.NotificationTemplateDto;
-import greencity.dto.PageableDto;
+import greencity.dto.*;
 import greencity.entity.enums.NotificationType;
 import greencity.entity.notifications.NotificationTemplate;
+import greencity.entity.order.Order;
 import greencity.entity.schedule.NotificationSchedule;
 import greencity.exceptions.NotFoundException;
 import greencity.repository.NotificationScheduleRepo;
 import greencity.repository.NotificationTemplateRepository;
+import greencity.repository.OrderRepository;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
@@ -25,10 +26,12 @@ import java.util.stream.Collectors;
 @Service
 @Transactional
 @AllArgsConstructor
-public class ManagementNotificationServiceImpl implements ManagementNotificationService {
+public class NotificationeServiceImpl implements NotificationeService {
     private NotificationTemplateRepository notificationTemplateRepository;
     private final ModelMapper modelMapper;
     private final NotificationScheduleRepo scheduleRepo;
+    private final OrderRepository orderRepository;
+    private final RestClient restClient;
 
     /**
      * {@inheritDoc}
@@ -83,5 +86,23 @@ public class ManagementNotificationServiceImpl implements ManagementNotification
     private NotificationScheduleDto getScheduleDto(NotificationTemplate notificationTemplate) {
         return modelMapper.map(scheduleRepo.getOne(
             notificationTemplate.getNotificationType()), NotificationScheduleDto.class);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void sendNotificationAboutViolation(AddingViolationsToUserDto dto, String language) {
+        Order order = orderRepository.findById(dto.getOrderID()).orElse(null);
+        UserViolationMailDto mailDto;
+        if (order != null) {
+            mailDto = UserViolationMailDto.builder()
+                .name(order.getUser().getRecipientName())
+                .email(order.getUser().getRecipientEmail())
+                .violationDescription(dto.getViolationDescription())
+                .language(language)
+                .build();
+            restClient.sendViolationOnMail(mailDto);
+        }
     }
 }
